refactor : QueueService 의 역할을 JobService 로 이동
